Refactor composer-app to use model app structure.
