+++ packages/apps/composer-app/src/containers/index.ts
@@ -1,0 +1,5 @@
+//
+// Copyright 2023 DXOS.org
+//
+
+export * from './ComposerDocument';
