+++ packages/apps/composer-app/src/layouts/SpaceLayout.tsx
