+++ packages/apps/composer-app/src/layouts/index.ts
@@ -1,0 +1,7 @@
+//
+// Copyright 2023 DXOS.org
+//
+
+export * from './AppLayout';
+export * from './SpaceLayout';
+export * from './SpaceSettingsLayout';
