+++ packages/apps/composer-app/src/pages/index.ts
@@ -3,3 +3,4 @@
 //
 
 export * from './SpacePage';
+export * from './SpacesPage';
