+++ packages/apps/tasks-app/src/components/index.ts
@@ -1,0 +1,8 @@
+//
+// Copyright 2023 DXOS.org
+//
+
+export * from './CheckboxItem';
+export * from './Input';
+export * from './List';
+export * from './Main';
